move appropriate rest tests from base to rest extention
